pyproject.toml
[project]
name = "product-horse"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "Max Caldwell", email = "max@scriptt.ai" }
]
dependencies = [
    "pydantic>=2.6.4",
    "llama-index>=0.10.26",
    "jupyterlab>=4.1.5",
    "nbdev>=2.3.13",
    "hypothesis>=6.99.9",
    "cloudflare-python>=1.0.1",
    "cloudflare>=2.19.2",
    "product-horse @ file:///Users/max/Documents/product_horse",
    "assemblyai>=0.23.1",
    "python-dotenv>=1.0.1",
    "testcontainers-postgres>=0.0.1rc1",
    "sqlmodel>=0.0.16",
    "openai>=1.14.3",
    "tiktoken>=0.6.0",
    "jsonschema>=4.21.1",
    "datasets>=2.18.0",
    "accelerate>=0.28.0",
    "llama-cpp-python>=0.2.57",
    "setuptools>=69.2.0",
    "instructor>=0.6.7",
    "anthropic>=0.20.0",
    "litellm>=1.34.5",
    "pandas>=2.2.1",
    "gradio>=4.23.0",
    "tqdm>=4.66.2",
    "llama-index-readers-database>=0.1.3",
    "llama-index-llms-anthropic>=0.1.7",
    "llama-index-vector-stores-chroma>=0.1.6",
    "moviepy>=1.0.3",
    "beautifulsoup4>=4.12.3",
    "seewav>=0.1.0",
    "scipy>=1.14.0",
    "opencv-python>=4.10.0.84",
    "pydub>=0.25.1",
    "pyav>=12.1.0",
    "multiprocess>=0.70.16",
    "strawberry-graphql[debug-server]>=0.235.1",
    "passlib>=1.7.4",
    "fastapi>=0.111.0",
    "pyjwt>=2.8.0",
    "openapi-python-client>=0.21.2",
    "certifi>=2024.7.4",
    "tenacity>=9.0.0",
]
readme = "README.md"
requires-python = ">= 3.11"

[tool.rye.scripts]
prep = 'nbdev_prepare'
test = 'nbdev_test'
export = 'nbdev_export'
graphql = 'strawberry server product_horse.graphql'
dev = 'rye run fastapi dev product_horse/graphql.py'
rls-local = '''psql -d product_horse -c "SET app.rls_user_password = '$RLS_USER_PASSWORD';" -f enable_rls.sql'''
workers = 'npx wrangler dev'
setup_d1_local = 'npx wrangler d1 execute ph-storage-db --file=/Users/max/Documents/product_horse/sql_files/setup_d1.sql'
setup_d1_remote = 'npx wrangler d1 execute ph-storage-db --file=/Users/max/Documents/product_horse/sql_files/setup_d1.sql --remote'
build_storage_client = 'rye run openapi-python-client generate --path /Users/max/Documents/product_horse/workers/openapi.yaml --output-path /Users/max/Documents/product_horse/storage_client --overwrite'
deploy_worker='npx wrangler deploy'

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["product_horse"]
